requirements.txt
# JyotiYantra Backend Dependencies
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
gunicorn==21.2.0
